Added a semaphore to control concurrency in batch_job_lambda.
