+++ dttools/src/Makefile
@@ -91,6 +91,7 @@ SOURCES = \
 	rmonitor_poll.c \
 	rmsummary.c \
 	set.c \
+    semaphore.c \
 	sha1.c \
 	shell.c \
 	sigdef.c \
